requirements.txt
python-dotenv==1.1.0
openai==1.82.0
chromadb==1.0.10